RDK-35523: Unit test improvements
Reason for change: Simplify DeviceIdentificationTest.
Move python script into Scripts, revert -j (freezes system),
fix getopts (had no effect), update README.
Test Procedure: Run Unit Tests.
Risks: Low
