+++ RdkServicesTest/README.md
@@ -5,22 +5,22 @@ Write unit tests for your plugin and add them to RdkServicesTest.
 Supply the CMake options to build your plugin in [build.sh](./Scripts/build.sh).
 Read the [GoogleTest User’s Guide](https://google.github.io/googletest/).
 
-## How to build ##
+## 1. Build ##
 
 ```shell script
 cd RdkServicesTest
-./Scripts/build.sh [-j<number>] [-D]
+./Scripts/build.sh [-D] [-t]
 
-where:
-    -j  number of threads (default maximum available threads)
-    -D  debug mode (default Release)"
+options:
+  -D                   Debug mode (default Release)
+  -t                   Build tests only (default everything)
 ```
 
 This builds Thunder, ThunderInterfaces, Rdkservices and installs them in thunder/install.
 
 Make sure you have packages _python3 python3-pip libsqlite3-dev libcurl4-openssl-dev_.
 
-## How to run ##
+## 2. Run ##
 
 ```shell script
 cd RdkServicesTest
@@ -31,7 +31,7 @@ This runs all tests, generates Valgrind report valgrind_log, and coverage info.
 
 Make sure you have package _valgrind_.
 
-## Coverage ##
+## 3. Coverage ##
 
 ```shell script
 cd RdkServicesTest
