+++ RdkServicesTest/Scripts/DeviceDiagnosticMock.py
